requirements.txt
google-generativeai
agentops